Fikser pre-query sjekk
